break down app.tsx into multiple pages
