+++ package.json
@@ -13,7 +13,8 @@
     "axios": "^1.7.2",
     "dotenv": "^16.4.5",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.24.1"
   },
   "devDependencies": {
     "@types/node": "^20.14.9",
